component_sharp_lr35902: Generator: Generalize the operand type-based switch/case
A really welcome improvement, since the complexity of the cases was combinatorial.

Also remove unnecessary `&` operators from the immediate variables in the execution method call.

As part of the simplification, for some instructions, both registers are mutable; this could be made precise, however, it would add complexity in an ugly form.